chore: Move extend function to function folder
Previously it was grouped with filters in lib/filters.
