+++ lib/create-nunjucks-environment.js
@@ -3,7 +3,7 @@ import nunjucks from 'nunjucks';
 import setAttribute from './filters/set-attribute.js';
 import setAttributes from './filters/set-attributes.js';
 import tojson from './filters/tojson.js';
-import extend from './filters/extend.js';
+import extend from './functions/extend.js';
 
 export default function createNunjucksEnvironment(loader) {
     const templatePaths = ['src'];
