+++ lib/functions/extend.js
@@ -5,7 +5,8 @@
  * with array.append(item), but not with any syntax that is available in both.
  *
  * Use:
- *     {% do extend(series, seriesItem) %}
+ *     {# Use `set` tag to avoid printing the return value of extend #}
+ *     {% set _ = extend(series, seriesItem) %}
  */
 export default function extend(value, element) {
     if (!Array.isArray(value)) {
