+++ lib/functions/extend.spec.js
